+++ appData/src/gb/include/projectiles.h
@@ -1,6 +1,8 @@
 #ifndef PROJECTILES_H
 #define PROJECTILES_H
 
+#include <gbdk/platform.h>
+
 #include "math.h"
 #include "collision.h"
 #include "gbs_types.h"
@@ -10,9 +12,9 @@
 
 extern projectile_def_t projectile_defs[MAX_PROJECTILES];
 
-void projectiles_init() BANKED;
-void projectiles_update() NONBANKED;
-void projectiles_render() NONBANKED;
+void projectiles_init(void) BANKED;
+void projectiles_update(void) NONBANKED;
+void projectiles_render(void) NONBANKED;
 
 #define PROJECTILE_ANIM_NOLOOP 0x01
 #define PROJECTILE_STRONG 0x02