+++ appData/src/gb/include/states/logo.h
@@ -1,9 +1,9 @@
 #ifndef STATE_LOGO_H
 #define STATE_LOGO_H
 
-#include <gb/gb.h>
+#include <gbdk/platform.h>
 
-void logo_init() BANKED;
-void logo_update() BANKED;
+void logo_init(void) BANKED;
+void logo_update(void) BANKED;
 
 #endif