+++ buildTools/darwin-x64/gbdk/examples/nes/smoke/smoketest.c
@@ -1,0 +1,8 @@
+#include <gbdk/platform.h>
+
+void main(void) {
+    DISPLAY_ON;
+    while(TRUE) {
+        vsync();
+    }
+}