+++ buildTools/win32-ia32/gbdk/examples/cross-platform/banks_autobank/src/srcfile_3.c
@@ -10,9 +10,9 @@ const uint8_t some_const_var_3 = 3;
 
 
 BANKREF(func_3)
-void func_3() BANKED
+void func_3(void) BANKED
 {
-    printf("Func3 in ROM bank %u\n", _current_bank);
+    printf("Func3 in ROM bank %u\n", CURRENT_BANK);
 }
 
 // A big constant array to take up space