+++ buildTools/win32-ia32/gbdk/include/gbdk/bcd.h
@@ -3,7 +3,7 @@
 
 #if defined(__TARGET_gb) || defined(__TARGET_ap) || defined(__TARGET_duck)
   #include <gb/bcd.h>
-#elif defined(__TARGET_sms) || defined(__TARGET_gg)
+#elif defined(__TARGET_sms) || defined(__TARGET_gg) || defined(__TARGET_msxdos)
   #error Not implemented yet
 #else
   #error Unrecognized port