+++ buildTools/win32-x64/gbdk/include/stdlib.h
@@ -6,7 +6,7 @@
 
 #include <types.h>
 
-#if !defined(__SDCC_mcs51) && !defined(__SDCC_ds390) && !defined(__SDCC_ds400) && !defined(__SDCC_hc08) && !defined(__SDCC_s08) && !defined(__SDCC_pic14) && !defined(__SDCC_pic16) && !defined(__SDCC_pdk13) && !defined(__SDCC_pdk14) && !defined(__SDCC_pdk15)
+#if !defined(__SDCC_mcs51) && !defined(__SDCC_ds390) && !defined(__SDCC_ds400) && !defined(__SDCC_hc08) && !defined(__SDCC_s08) && !defined(__SDCC_mos6502) && !defined(__SDCC_mos65c02) && !defined(__SDCC_pic14) && !defined(__SDCC_pic16) && !defined(__SDCC_pdk13) && !defined(__SDCC_pdk14) && !defined(__SDCC_pdk15)
 #define __reentrant
 #endif
 
@@ -14,12 +14,12 @@
     returned to the parent.
     All open streams are flushed and closed.
 */
-void exit(int status);
+void exit(int status) OLDCALL;
 
 #if 0
 /** Compatibility function.  Not implemented.
  */
-int getkey(void);
+int getkey(void) OLDCALL;
 #endif
 
 /** Returns the absolute value of int __i__
@@ -69,6 +69,9 @@ long atol(const char *s);
     @param radix  Numerical base for converted number, ex: 10 is decimal base
                   (parameter is required but not utilized on Game Boy and Analogue Pocket)
 
+    Can be used with @ref set_bkg_based_tiles() for printing if
+    the digit character tiles are not ascii-mapped.
+
     Returns:    Pointer to converted string
  */
 char *itoa(int n, char *s, unsigned char radix) OLDCALL;
@@ -78,6 +81,9 @@ char *itoa(int n, char *s, unsigned char radix) OLDCALL;
     @param s      String to store the converted number
     @param radix  Numerical base for converted number, ex: 10 is decimal base
                   (parameter is required but not utilized on Game Boy and Analogue Pocket)
+
+    Can be used with @ref set_bkg_based_tiles() for printing if
+    the digit character tiles are not ascii-mapped.
 
     Returns:    Pointer to converted string
  */
@@ -89,6 +95,9 @@ char *uitoa(unsigned int n, char *s, unsigned char radix) OLDCALL;
     @param radix  Numerical base for converted number, ex: 10 is decimal base
                   (parameter is required but not utilized on Game Boy and Analogue Pocket)
 
+    Can be used with @ref set_bkg_based_tiles() for printing if
+    the digit character tiles are not ascii-mapped.
+
     Returns:    Pointer to converted string
  */
 char *ltoa(long n, char *s, unsigned char radix) OLDCALL;
@@ -98,6 +107,9 @@ char *ltoa(long n, char *s, unsigned char radix) OLDCALL;
     @param s      String to store the converted number
     @param radix  Numerical base for converted number, ex: 10 is decimal base
                   (parameter is required but not utilized on Game Boy and Analogue Pocket)
+
+    Can be used with @ref set_bkg_based_tiles() for printing if
+    the digit character tiles are not ascii-mapped.
 
     Returns:    Pointer to converted string
  */