+++ src/components/music/SongTracker.tsx
@@ -1,4 +1,10 @@
-import React, { useState, useRef, useEffect, useCallback } from "react";
+import React, {
+  useState,
+  useRef,
+  useEffect,
+  useCallback,
+  useMemo,
+} from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 import { PatternCell } from "lib/helpers/uge/song/PatternCell";
@@ -10,18 +16,44 @@ import { SequenceEditor } from "./SequenceEditor";
 import { SongRow } from "./SongRow";
 import scrollIntoView from "scroll-into-view-if-needed";
 import { SongGridHeaderCell } from "./SongGridHeaderCell";
-import { ipcRenderer } from "electron";
+import { IpcRendererEvent, ipcRenderer } from "electron";
+import { getInstrumentTypeByChannel, getInstrumentListByType } from "./helpers";
 import {
-  getInstrumentTypeByChannel,
-  getInstrumentListByType,
   parseClipboardToPattern,
   parsePatternFieldsToClipboard,
-} from "./helpers";
+} from "./musicClipboardHelpers";
 import { getKeys, KeyWhen } from "lib/keybindings/keyBindings";
 import trackerActions from "store/features/tracker/trackerActions";
 import clipboardActions from "store/features/clipboard/clipboardActions";
 import { clipboard } from "store/features/clipboard/clipboardHelpers";
 import { clamp, cloneDeep, mergeWith } from "lodash";
+
+function getSelectedTrackerFields(
+  selectionRect: SelectionRect | undefined,
+  selectionOrigin: Position | undefined
+) {
+  const selectedTrackerFields = [];
+  if (selectionRect) {
+    for (
+      let i = selectionRect.x;
+      i <= selectionRect.x + selectionRect.width;
+      i++
+    ) {
+      for (
+        let j = selectionRect.y;
+        j <= selectionRect.y + selectionRect.height;
+        j++
+      ) {
+        selectedTrackerFields.push(j * ROW_SIZE + i);
+      }
+    }
+  } else if (selectionOrigin) {
+    selectedTrackerFields.push(
+      selectionOrigin.y * ROW_SIZE + selectionOrigin.x
+    );
+  }
+  return selectedTrackerFields;
+}
 
 interface SongTrackerProps {
   sequenceId: number;
@@ -85,12 +117,13 @@ export const SongTracker = ({
   const startPlaybackPosition = useSelector(
     (state: RootState) => state.tracker.startPlaybackPosition
   );
+  const subpatternEditorFocus = useSelector(
+    (state: RootState) => state.tracker.subpatternEditorFocus
+  );
 
   const patternId = song?.sequence[sequenceId] || 0;
   const pattern = song?.patterns[patternId];
 
-  const [selectedTrackerFields, setSelectedTrackerFields] =
-    useState<number[]>();
   const [selectionOrigin, setSelectionOrigin] =
     useState<Position | undefined>();
   const [selectionRect, setSelectionRect] =
@@ -98,38 +131,13 @@ export const SongTracker = ({
   const [isSelecting, setIsSelecting] = useState(false);
   const [isMouseDown, setIsMouseDown] = useState(false);
 
-  useEffect(() => {
-    const newSelectedTrackerFields = [];
-    if (selectionRect) {
-      for (
-        let i = selectionRect.x;
-        i <= selectionRect.x + selectionRect.width;
-        i++
-      ) {
-        for (
-          let j = selectionRect.y;
-          j <= selectionRect.y + selectionRect.height;
-          j++
-        ) {
-          newSelectedTrackerFields.push(j * ROW_SIZE + i);
-        }
-      }
-    } else if (selectionOrigin) {
-      newSelectedTrackerFields.push(
-        selectionOrigin.y * ROW_SIZE + selectionOrigin.x
-      );
-    }
-    setSelectedTrackerFields(newSelectedTrackerFields);
-  }, [selectionOrigin, selectionRect]);
-
-  const [selectedTrackerRows, setSelectedTrackerRows] = useState<number[]>();
-  useEffect(() => {
-    const newSelectedTrackerRows = selectedTrackerFields?.map((f) =>
-      Math.floor(f / ROW_SIZE)
-    );
-
-    setSelectedTrackerRows(newSelectedTrackerRows);
-  }, [selectedTrackerFields]);
+  const selectedTrackerFields = useMemo(
+    () => getSelectedTrackerFields(selectionRect, selectionOrigin),
+    [selectionOrigin, selectionRect]
+  );
+  const selectedTrackerRows = selectedTrackerFields?.map((f) =>
+    Math.floor(f / ROW_SIZE)
+  );
 
   const [playbackState, setPlaybackState] = useState([0, 0]);
 
@@ -137,7 +145,7 @@ export const SongTracker = ({
     setPlaybackState(startPlaybackPosition);
   }, [setPlaybackState, startPlaybackPosition]);
   useEffect(() => {
-    const listener = (_event: any, d: any) => {
+    const listener = (_event: IpcRendererEvent, d: any) => {
       if (d.action === "update") {
         setPlaybackState(d.update);
       }
@@ -154,46 +162,36 @@ export const SongTracker = ({
     (state: RootState) => state.tracker.selectedChannel
   );
 
-  useEffect(() => {
-    if (activeField !== undefined) {
-      const newChannelId = Math.floor(
-        (activeField % ROW_SIZE) / CHANNEL_FIELDS
+  if (activeField !== undefined) {
+    const newChannelId = Math.floor((activeField % ROW_SIZE) / CHANNEL_FIELDS);
+    dispatch(trackerActions.setSelectedChannel(newChannelId));
+    if (activeField % CHANNEL_FIELDS >= 2) {
+      dispatch(
+        trackerActions.setSelectedEffectCell(Math.floor(activeField / ROW_SIZE))
       );
-      dispatch(trackerActions.setSelectedChannel(newChannelId));
-      if (activeField % CHANNEL_FIELDS >= 2) {
-        dispatch(
-          trackerActions.setSelectedEffectCell(
-            Math.floor(activeField / ROW_SIZE)
-          )
-        );
-      }
     }
-  }, [activeField, dispatch]);
+  }
 
   const playingRowRef = useRef<HTMLSpanElement>(null);
-  useEffect(() => {
-    if (playingRowRef && playingRowRef.current) {
-      if (playing) {
-        playingRowRef.current.scrollIntoView({
-          behavior: "auto",
-          block: "center",
-          inline: "nearest",
-        });
-      }
+  if (playingRowRef && playingRowRef.current) {
+    if (playing) {
+      playingRowRef.current.scrollIntoView({
+        behavior: "auto",
+        block: "center",
+        inline: "nearest",
+      });
     }
-  }, [playing, playbackState]);
+  }
 
   const activeFieldRef = useRef<HTMLSpanElement>(null);
-  useEffect(() => {
-    if (activeFieldRef && activeFieldRef.current) {
-      if (!playing) {
-        scrollIntoView(activeFieldRef.current.parentElement as Element, {
-          scrollMode: "if-needed",
-          block: "nearest",
-        });
-      }
+  if (activeFieldRef && activeFieldRef.current) {
+    if (!playing) {
+      scrollIntoView(activeFieldRef.current.parentElement as Element, {
+        scrollMode: "if-needed",
+        block: "nearest",
+      });
     }
-  }, [playing, activeField]);
+  }
 
   const transposeSelectedTrackerFields = useCallback(
     (change: number, large: boolean) => {
@@ -286,7 +284,7 @@ export const SongTracker = ({
 
   const insertTrackerFields = useCallback(
     (uninsert: boolean) => {
-      if (pattern && activeField) {
+      if (pattern && activeField !== undefined) {
         const newChannelId = Math.floor(
           (activeField % ROW_SIZE) / CHANNEL_FIELDS
         );
@@ -301,12 +299,7 @@ export const SongTracker = ({
             newPattern[i][newChannelId] = newPattern[i - 1][newChannelId];
           }
         }
-        newPattern[uninsert ? 63 : startRow][newChannelId] = {
-          note: null,
-          instrument: null,
-          effectcode: null,
-          effectparam: null,
-        };
+        newPattern[uninsert ? 63 : startRow][newChannelId] = new PatternCell();
         dispatch(
           trackerDocumentActions.editPattern({
             patternId: patternId,
@@ -319,7 +312,11 @@ export const SongTracker = ({
   );
 
   const handleMouseDown = useCallback(
-    (e: any) => {
+    (e: MouseEvent) => {
+      if (!e.target || !(e.target instanceof HTMLElement)) {
+        return;
+      }
+
       const fieldId = e.target.dataset["fieldid"];
       const rowId = e.target.dataset["row"];
 
@@ -369,7 +366,10 @@ export const SongTracker = ({
   );
 
   const handleMouseUp = useCallback(
-    (_e: any) => {
+    (e: MouseEvent) => {
+      if (!e.target || !(e.target instanceof HTMLElement)) {
+        return;
+      }
       if (isMouseDown) {
         setIsMouseDown(false);
       }
@@ -378,7 +378,10 @@ export const SongTracker = ({
   );
 
   const handleMouseMove = useCallback(
-    (e: any) => {
+    (e: MouseEvent) => {
+      if (!e.target || !(e.target instanceof HTMLElement)) {
+        return;
+      }
       if (isMouseDown) {
         const fieldId = e.target.dataset["fieldid"];
 
@@ -402,7 +405,7 @@ export const SongTracker = ({
     [isMouseDown, selectionOrigin]
   );
 
-  const handleKeys = useCallback(
+  const handleKeyDown = useCallback(
     (e: KeyboardEvent) => {
       const editPatternCell =
         (type: keyof PatternCell) => (value: number | null) => {
@@ -487,12 +490,12 @@ export const SongTracker = ({
 
       if (e.key === "Escape") {
         e.preventDefault();
-        setSelectedTrackerFields(undefined);
         setSelectionOrigin(undefined);
+        setSelectionRect(undefined);
       }
 
       if (e.key === "Backspace" || e.key === "Delete") {
-        if ((e.shiftKey || e.ctrlKey) && activeField) {
+        if ((e.shiftKey || e.ctrlKey) && activeField !== undefined) {
           e.preventDefault();
           insertTrackerFields(true);
           return;
@@ -505,7 +508,7 @@ export const SongTracker = ({
       }
 
       if (e.key === "Insert" || e.key === "Enter") {
-        if (activeField) {
+        if (activeField !== undefined) {
           insertTrackerFields(false);
           return;
         }
@@ -633,20 +636,14 @@ export const SongTracker = ({
     ]
   );
 
-  const handleKeysUp = useCallback(
-    (e: KeyboardEvent) => {
-      if (activeField) {
-        // console.log(e.key);
-      }
-      if (!e.shiftKey) {
-        setIsSelecting(false);
-      }
-    },
-    [activeField]
-  );
+  const handleKeyUp = useCallback((e: KeyboardEvent) => {
+    if (!e.shiftKey) {
+      setIsSelecting(false);
+    }
+  }, []);
 
   const handleWheel = useCallback(
-    (e: any) => {
+    (e: WheelEvent) => {
       if (e.ctrlKey) {
         e.preventDefault();
         const delta = e.deltaY === 0 ? e.deltaX : e.deltaY;
@@ -694,46 +691,53 @@ export const SongTracker = ({
   );
 
   useEffect(() => {
-    window.addEventListener("keydown", handleKeys);
-    window.addEventListener("keyup", handleKeysUp);
+    window.addEventListener("keydown", handleKeyDown);
+    window.addEventListener("keyup", handleKeyUp);
     window.addEventListener("mousedown", handleMouseDown);
     window.addEventListener("mouseup", handleMouseUp);
     window.addEventListener("mousemove", handleMouseMove);
     window.addEventListener("wheel", handleWheel, { passive: false });
 
     return () => {
-      window.removeEventListener("keydown", handleKeys);
-      window.removeEventListener("keyup", handleKeysUp);
+      window.removeEventListener("keydown", handleKeyDown);
+      window.removeEventListener("keyup", handleKeyUp);
       window.removeEventListener("mousedown", handleMouseDown);
       window.removeEventListener("mouseup", handleMouseUp);
       window.removeEventListener("mousemove", handleMouseMove);
       window.removeEventListener("wheel", handleWheel);
     };
-  });
+  }, [
+    handleKeyDown,
+    handleKeyUp,
+    handleMouseDown,
+    handleMouseMove,
+    handleMouseUp,
+    handleWheel,
+  ]);
 
   useEffect(() => {
-    document.addEventListener("selectionchange", onSelectAll);
-
-    return () => {
-      document.removeEventListener("selectionchange", onSelectAll);
-    };
-  });
+    if (!subpatternEditorFocus) {
+      document.addEventListener("selectionchange", onSelectAll);
+      return () => {
+        document.removeEventListener("selectionchange", onSelectAll);
+      };
+    }
+  }, [onSelectAll, subpatternEditorFocus]);
 
   const onFocus = useCallback(
     (_e: React.FocusEvent<HTMLDivElement>) => {
-      if (!activeField) {
+      if (activeField === undefined) {
         setActiveField(0);
       }
     },
     [activeField, setActiveField]
   );
 
-  const onBlur = useCallback(
-    (_e: React.FocusEvent<HTMLDivElement>) => {
-      setActiveField(undefined);
-    },
-    [setActiveField]
-  );
+  const onBlur = useCallback((_e: React.FocusEvent<HTMLDivElement>) => {
+    setActiveField(undefined);
+    setSelectionOrigin(undefined);
+    setSelectionRect(undefined);
+  }, []);
 
   const onCopy = useCallback(() => {
     if (pattern && selectedTrackerFields) {
@@ -797,15 +801,17 @@ export const SongTracker = ({
 
   // Clipboard
   useEffect(() => {
-    window.addEventListener("copy", onCopy);
-    window.addEventListener("cut", onCut);
-    window.addEventListener("paste", onPaste);
-    return () => {
-      window.removeEventListener("copy", onCopy);
-      window.removeEventListener("cut", onCut);
-      window.removeEventListener("paste", onPaste);
-    };
-  }, [onCopy, onCut, onPaste]);
+    if (!subpatternEditorFocus) {
+      window.addEventListener("copy", onCopy);
+      window.addEventListener("cut", onCut);
+      window.addEventListener("paste", onPaste);
+      return () => {
+        window.removeEventListener("copy", onCopy);
+        window.removeEventListener("cut", onCut);
+        window.removeEventListener("paste", onPaste);
+      };
+    }
+  }, [onCopy, onCut, onPaste, subpatternEditorFocus]);
 
   return (
     <div